polymarket-agent-chat/backend/requirements.txt
fastapi
uvicorn[standard]
httpx
anthropic
liquidtrading-python
py-clob-client
python-dotenv
